exp [all random seeds] test_kmeans_plusplus_norms test_kmeans_elkan_results
