+++ azure-pipelines.yml
@@ -13,19 +13,7 @@ jobs:
   pool:
     vmImage: ubuntu-20.04
   steps:
-    - bash: |
-        set -ex
-        if [[ $BUILD_REASON == "PullRequest" ]]; then
-          # By default pull requests use refs/pull/PULL_ID/merge as the source branch
-          # which has a "Merge ID into ID" as a commit message. The latest commit
-          # message is the second to last commit
-          COMMIT_ID=$(echo $BUILD_SOURCEVERSIONMESSAGE | awk '{print $2}')
-          message=$(git log $COMMIT_ID -1 --pretty=%B)
-          message=$(echo $message)
-        else
-          message=$BUILD_SOURCEVERSIONMESSAGE
-        fi
-        echo "##vso[task.setvariable variable=message;isOutput=true]$message"
+    - script: python build_tools/azure/get_commit_message.py
       name: commit
       displayName: Get source version message
